# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SNP - ĐỒNG BỘ DỮ LIỆU",
    "version": "1.0.0",
    "description": "Hệ thống đồng bộ dữ liệu EOffice &mdash; SNP"
  },
  "servers": [
    {
      "url": "http://localhost:3025/api"
    },
    {
      "url": "http://SNP-DongBoDuLieu:3025/api"
    }
  ],
  "paths": {
...
    "/sync-passport/migrate/test-get-list": {
      "post": {
        "tags": [
          "Sync Passport"
        ],
        "summary": "Thử nghiệm lấy danh sách phiếu mượn hộ chiếu cần đồng bộ.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lastSyncTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Thời điểm đồng bộ cuối (mặc định 1970)."
                  },
                  "syncJobId": {
                    "type": "string",
                    "description": "ID của job đồng bộ hiện tại."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Danh sách các phiếu mượn hộ chiếu tìm được."
          }
        }
      }
    },
    "/sync-passport/migrate/test-process-one": {
      "post": {
        "tags": [
          "Sync Passport"
        ],
        "summary": "Thử nghiệm xử lý (đồng bộ) một bản ghi phiếu mượn hộ chiếu.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "syncJobId"
                ],
                "properties": {
                  "syncJobId": {
                    "type": "string",
                    "description": "ID của bản ghi/ID Job để xử lý."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kết quả xử lý bản ghi."
          }
        }
      }
    },
...
  },
...
}